refactor(ff-preview): split player.rs into player, player_handle, and player_runner modules
