+++ crates/ff-preview/src/playback/mod.rs
@@ -9,7 +9,9 @@
 //! | `master_clock`     | Internal `MasterClock` |
 //! | `sink`             | [`FrameSink`] trait, [`RgbaFrame`], [`RgbaSink`] |
 //! | `decode_buffer`    | [`DecodeBuffer`], [`FrameResult`], [`SeekEvent`] |
-//! | `player`           | [`PreviewPlayer`] |
+//! | `player`           | [`PreviewPlayer`], [`PlayerCommand`] |
+//! | `player_handle`    | [`PlayerHandle`] |
+//! | `player_runner`    | [`PlayerRunner`] |
 //! | `async_player`     | [`AsyncPreviewPlayer`] (tokio feature) |
 //! | `playback_inner`   | Unsafe `FFmpeg` calls (`SwsRgbaConverter`, PCM extraction) |
 
@@ -22,6 +24,8 @@ pub(crate) mod clock;
 pub(crate) mod decode_buffer;
 pub(crate) mod master_clock;
 pub(crate) mod player;
+pub(crate) mod player_handle;
+pub(crate) mod player_runner;
 pub(crate) mod sink;
 
 #[cfg(feature = "tokio")]
@@ -29,7 +33,9 @@ pub(crate) mod async_player;
 
 pub use clock::PlaybackClock;
 pub use decode_buffer::{DecodeBuffer, DecodeBufferBuilder, FrameResult, SeekEvent};
-pub use player::{PlayerCommand, PlayerHandle, PlayerRunner, PreviewPlayer};
+pub use player::{PlayerCommand, PreviewPlayer};
+pub use player_handle::PlayerHandle;
+pub use player_runner::PlayerRunner;
 pub use sink::{FrameSink, RgbaFrame, RgbaSink};
 
 #[cfg(feature = "tokio")]
