+++ crates/ff-preview/src/timeline/mod.rs
@@ -38,7 +38,8 @@ use crate::event::PlayerEvent;
 use crate::playback::SwsRgbaConverter;
 use crate::playback::decode_buffer::{DecodeBuffer, FrameResult};
 use crate::playback::master_clock::MasterClock;
-use crate::playback::player::{PlayerCommand, PlayerHandle};
+use crate::playback::player::PlayerCommand;
+use crate::playback::player_handle::PlayerHandle;
 use crate::playback::sink::FrameSink;
 
 // ── Constants ─────────────────────────────────────────────────────────────────
